Renamed Config to ConfigService (#7) (#7)
